% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about_data.R
\name{about_data}
\alias{about_data}
\title{Returns the Number of Observations and Variables in a Data Frame}
\usage{
about_data(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A character string
}
\description{
Sometimes you will want to post documents (e.g., R markdown) in a public
 forum (e.g., GitHub), but you don't want to reveal any raw data values.
 In such a case, using str() or glimpse() may be problematic. about_data()
 is a simple function that returns the number of observations and variables
 in a data frame, but none of the data values.
}
\examples{
data("mtcars")
about_data(mtcars)
names(mtcars)
}
