% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_na.R
\name{all_na}
\alias{all_na}
\title{Find rows that contain only NA values}
\usage{
all_na(x)
}
\arguments{
\item{x}{A data frame}
}
\description{
Given a data frame, this function returns the row number(s) where the value
 for every variable(s) in the row(s) is/are NA.
}
\examples{
df <- data.frame(
var1 = c(1, NA, 2, NA, 5, NA),
var2 = c(NA, "b", "c", NA, NA, NA),
var3 = c(TRUE, FALSE, TRUE, NA, FALSE, NA)
)

all_na(df)
}
