% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_catvars.R
\name{check_catvars}
\alias{check_catvars}
\title{Quick Look at Categorical Variables}
\usage{
check_catvars(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A list of frequency tables for all categorical variables in x.
}
\description{
Sometimes when cleaning data, we want to see all the levels that exist for
all categorical variables.
}
\examples{
# Load mtcars
data(mtcars)

# Check a single categorical variable
check_catvars(mtcars$cyl)

# Check every variable in a data frame
lapply(mtcars, check_catvars)
}
