% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Automate creation of a data codebook}
\usage{
codebook(x, label = NA)
}
\arguments{
\item{x}{A data frame}

\item{label}{Optional vector of variable labels. Default is NA.}
}
\value{
A list of lists
}
\description{
The codebook function assists with the creation of a codebook for a given
 data frame.
}
\examples{
data("mtcars")

# Optional: Create label vector
labels <- c("Miles Per Gallon", "Number of cylinders")

# Codebook with one variable
codebook(mtcars["mpg"], label = labels)

# Codebook with multiple variables
varlist <- c("mpg", "cyl")
codebook(mtcars[varlist], label = labels)

# Codebook with all variables
codebook(mtcars)

# Do not use:
# lapply(mtcars[varlist], codebook)
}
