% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_get_df_attributes.R
\name{codebook_get_df_attributes}
\alias{codebook_get_df_attributes}
\title{Get Data Frame Attributes}
\usage{
codebook_get_df_attributes(df, path = NA)
}
\arguments{
\item{df}{The saved file at `path`, read into memory as a data frame}

\item{path}{The path to the saved dataset of interest}
}
\value{
A tibble of selected data frame attributes
}
\description{
Codebook expects that `df ` is a data frame that you have read into memory
from a saved data file. Please provide the path to the saved data file. This
function gets selected attributes about file saved at `path` and stores
those attributes in a data frame, which is later turned into a flextable and
added to the codebook document.
}
