% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_matrix.R
\name{conf_matrix}
\alias{conf_matrix}
\title{Confusion Matrix}
\usage{
conf_matrix(truth, prediction, show_matrix = TRUE)
}
\arguments{
\item{truth}{A vector containing actual or "gold standard" outcomes.}

\item{prediction}{A vector containing predicted values or observed
diagnostic values.}

\item{show_matrix}{A logical (TRUE or FALSE). Set to TRUE by default. Change
to FALSE if you want a confusion matrix table to be printed to the console.}
}
\description{
Create a confusion matrix along with the following measures
 of performance
}
\examples{
data("mtcars")

# Create "true" outcome vector
truth <- as.integer(mtcars$vs)

# Create a vector meant to represent a set of predicted outcomes
set.seed(123)
predicted <- sample(0:1, length(truth), replace = TRUE)

# Create confusion matrix
my_cm <- conf_matrix(truth = truth, prediction = predicted)
my_cm

# --------------------------------------------------------------------------
# Example of running conf_matrix over multiple variables
df <- data.frame(
x1 = factor(c("No", "Yes")),
x2 = factor(c("No", "Yes")),
x3 = factor(c("Yes", "No"))
)
# Calculate measures
test <- conf_matrix(truth = df$x1, prediction = df$x2, show_matrix = FALSE)
as.data.frame(test)

# Make empty table
results <- data.frame(
  var = NA,
  tp = NA,
  fp = NA,
  fn = NA,
  tn = NA,
  sensativity = NA,
  specificity = NA,
  fpr = NA,
  fnr = NA,
  accuracy = NA,
  misclassification = NA,
  precision = NA,
  prevalence = NA
)

# for each screener item:
#   capture the name of the screener item
#   put the name into the first column of the results df
#   capture the performance measures of the screener item
#   put each performance measure into the coresponding element in the results df
var <- names(df[2:3])
r <- 1
for (i in var) {
  results[r, 1] <- i
  results[r, 2:13] <- conf_matrix(truth = df$x1, prediction = df[[i]], show_matrix = FALSE)
  r <- r + 1
}
}
\references{
Wikipedia - Sensativity and Specificity
 \url{https://en.wikipedia.org/wiki/Sensitivity_and_specificity}
}
