% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ids.R
\name{count_ids}
\alias{count_ids}
\title{Count the number of unique ID's}
\usage{
count_ids(x, description = "values", ...)
}
\arguments{
\item{x}{A variable (vector). Typically an ID variable.}

\item{description}{A character string that describes the unit of observation
in the return value (e.g., "women", "participants", "schools", "cities").
The default is "values".}

\item{...}{Additional arguments}
}
\value{
A character string
}
\description{
This function returns a value representing the number of unique
  values for a given variable as a nicely formatted value. In input variable
  could be any variable, however, the intent is for it to be a variable that
  identifies each unit of observation in the data (ID variable).
}
\examples{
df <- tibble::tibble(
  id = rep(1:5, 2),
  x = rnorm(10)
)

count_ids(df$id)
}
