% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{Estimate Percents and 95 Percent Confidence Intervals in dplyr Pipelines}
\usage{
freq_table(x, t_prob = 0.975, ci_type = "log", ...)
}
\arguments{
\item{x}{A grouped tibble, i.e., class == "grouped_df"}

\item{t_prob}{(1 - alpha / 2). Default value is 0.975, which corresponds to
an alpha of 0.05. Used to calculate a critical value from Student's t
distribution with n - 1 degrees of freedom.}

\item{ci_type}{Selects the method used to estimate 95 percent confidence intervals.
The default for one-way and two-way tables is log transformed ("log"). For
one-way tables only, ci_type can optionally calculate Wald ("linear")
confidence intervals using the "wald" argument.}

\item{...}{Other parameters to be passed on.}
}
\value{
A tibble
}
\description{
The freq_table function produces one-way and two-way frequency
  tables for categorical variables. In addition to frequencies, the
  freq_table function displays percentages, and the standard errors and 95%
  confidence intervals of the percentages. For two-way tables only,
  freq_table also displays row (subgroup) percentages, standard errors,
  and 95 percent confidence intervals.

  freq_table is intended to be used in a dplyr pipeline. Specifically,
  freq_table expects the x argument to be a grouped tibble created with
  dplyr's group_by function.

  All standard errors are calculated as some version of:
  sqrt(proportion * (1 - proportion) / (n - 1))

  For one-way tables, the default 95 percent confidence intervals displayed are
  log transformed confidence intervals equivalent to those used by Stata.
  Additionally, freq_table will return Wald ("linear") confidence intervals
  if the argument to ci_type = "wald".

  For two-way tables, freq_table returns log transformed confidence
  intervals equivalent to those used by Stata.
}
\examples{
library(tidyverse)

data(mtcars)

# One-way frequency table

mtcars \%>\%
  group_by(am) \%>\%
  freq_table()

#> # A tibble: 2 x 8
#>      am     n n_total percent       se   t_crit  lcl_log  ucl_log
#>     <dbl> <int>   <int>   <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#> 1     0    19      32  59.375 8.820997 2.039513 40.94225 75.49765
#> 2     1    13      32  40.625 8.820997 2.039513 24.50235 59.05775

# Two-way frequency table

mtcars \%>\%
  group_by(am, cyl) \%>\%
  freq_table()

#> A tibble: 6 x 13
#>      am   cyl     n n_group n_total percent_total se_total lcl_total_log ucl_total_log percent_row
#>     <dbl> <dbl> <int>   <int>   <int>         <dbl>    <dbl>         <dbl>         <dbl>       <dbl>
#> 1     0     4     3      19      32         9.375 5.235146      2.859820      26.65944    15.78947
#> 2     0     6     4      19      32        12.500 5.939887      4.506576      30.18928    21.05263
#> 3     0     8    12      19      32        37.500 8.695104     21.969117      56.11463    63.15789
#> 4     1     4     8      13      32        25.000 7.777138     12.514741      43.71685    61.53846
#> 5     1     6     3      13      32         9.375 5.235146      2.859820      26.65944    23.07692
#> 6     1     8     2      13      32         6.250 4.347552      1.446671      23.24074    15.38462
#> # ... with 3 more variables: se_row <dbl>, lcl_row_log <dbl>, ucl_row_log <dbl>
}
\references{
Agresti, A. (2012). Categorical Data Analysis (3rd ed.). Hoboken, NJ: Wiley.

  SAS documentation: https://support.sas.com/documentation/cdl/en/statug/63347/HTML/default/viewer.htm#statug_surveyfreq_a0000000221.htm

  Stata documentation: https://www.stata.com/manuals13/rproportion.pdf
}
