% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_test.R
\name{freq_test}
\alias{freq_test}
\alias{freq_test.freq_table_one_way}
\alias{freq_test.freq_table_two_way}
\title{Hypothesis Testing For Frequency Tables}
\usage{
freq_test(x, ...)

\method{freq_test}{freq_table_one_way}(x, ...)

\method{freq_test}{freq_table_two_way}(x, method = "pearson", ...)
}
\arguments{
\item{x}{A tibble of class freq_table_one_way or freq_table_two_way.}

\item{...}{Other parameters to be passed on.}

\item{method}{Options for this parameter control the method used to
  calculate p-values.

  For the freq_table_two_way class, the options are "pearson" (default) -
  to use Pearson's chi-square test of independence, and "fisher" - to use
  Fisher's exact test.}
}
\value{
A tibble.
}
\description{
The freq_test function is an S3 generic. It currently has
  methods for conducting hypothesis tests on one-way and two-way frequency
  tables. Further, it is made to work in a dplyr pipeline with the
  freq_table function.
}
\examples{
library(tidyverse)
library(bfuncs)

data(mtcars)

# Test equality of proportions

mtcars \%>\%
  group_by(am) \%>\%
  freq_table() \%>\%
  freq_test() \%>\%
  select(1:6, p_chi2_pearson)

#>  # A tibble: 2 x 7
#>     am     n n_total percent   lcl   ucl p_chi2_pearson
#>  <dbl> <int>   <int>   <dbl> <dbl> <dbl>          <dbl>
#>  1   0    19      32   59.38 40.94 75.50      0.2888444
#>  2   1    13      32   40.62 24.50 59.06      0.2888444

# Chi-square test of independence

mtcars \%>\%
  group_by(am, vs) \%>\%
  freq_table() \%>\%
  freq_test() \%>\%
  select(1:8, p_chi2_pearson)

#> # A tibble: 4 x 9
#>      am    vs     n n_row n_total percent_row lcl_row ucl_row p_chi2_pearson
#>   <dbl> <dbl> <int> <int>   <int>       <dbl>   <dbl>   <dbl>          <dbl>
#> 1     0     0    12    19      32       63.16   38.76   82.28      0.3409429
#> 2     0     1     7    19      32       36.84   17.72   61.24      0.3409429
#> 3     1     0     6    13      32       46.15   20.83   73.63      0.3409429
#> 4     1     1     7    13      32       53.85   26.37   79.17      0.3409429
}
