% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n.R
\name{get_n}
\alias{get_n}
\title{Formatted Group Sample Size for Tables}
\usage{
get_n(.data, ...)
}
\arguments{
\item{.data}{A data frame or tibble}

\item{...}{A dplyr::filter expression. Used to select subgroup.}
}
\value{
A character string
}
\description{
Given a tibble and a filter expression, get_n returns
  the group sample size formatted as "N = XXXX". Made to work in a dplyr
  pipeline, and used when creating tables for publications / reports.
}
\examples{
library(tidyverse)
library(bfuncs)

data(mtcars)

# Get sample size for cars with 4 cylinders
mtcars \%>\% get_n(cyl == 4)

#> [1] "N = 11"
}
