% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean95.R
\name{mean95}
\alias{mean95}
\title{Calculate Means and 95 Percent Confidence Intervals}
\usage{
mean95(x, digits = 3)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{Rounds the values returned to the specified number of decimal
places (default 3).}
}
\value{
A named vector containing the mean, sd, and 95% confidence interval.
}
\description{
The mean95 function calculates the mean, sd, and 95% confidence interval for
a numeric vector.
}
\examples{
data("mtcars")

# Single univariate analysis
mean95(mtcars$mpg)
with(mtcars, mean95(mpg))

# Single bivariate analyses
with(mtcars, by(mpg, am, mean95))
}
