% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_center.R
\name{mean_center}
\alias{mean_center}
\title{Mean Center Numeric Variables}
\usage{
mean_center(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector containing x minus the mean of x. Note that NA's are ignored.
}
\description{
Function to mean center predictor variables.
}
\examples{
# load mtcars data
data(mtcars)

# Mean center a single numeric variable
(mtcars$c_mpg <- mean_center(mtcars$mpg))

# Mean center multiple numeric variables
centered <- do.call(cbind, lapply(mtcars[c("mpg", "hp")], mean_center))
colnames(centered) <- c("c_mpg", "c_hp")
mtcars2 <- cbind(mtcars, centered)
}
