% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/props.R
\name{props}
\alias{props}
\title{Estimate Percents and 95 Percent Confidence Intervals}
\usage{
props(x, citype = "normal", digits = 5)
}
\arguments{
\item{x}{A vector or a list element. Should be a categorical variable.}

\item{citype}{Choose a method for calculating confidence intervals. Either
"normal" or "logit" (Agresti, 2012).}

\item{digits}{Number of digits after the decimal point for percents.s}
}
\value{
A data frame containing the variable name and unique levels
 (categories). Within each level, the number of non-missing values,
 percent, and 95 percent confidence interval for the percent are returned.
}
\description{
The props function calculates the number of nonmissing values,
 percent, and 95 percent confidence interval for categorical variables.
}
\examples{
# data("mtcars")

# Single univariate analysis
props(mtcars$cyl, citype = "normal", digits = 2)
with(mtcars, props(cyl))
}
\references{
Agresti, A. (2012). Categorical Data Analysis (3rd ed.). Hoboken, NJ: Wiley.
}
