% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_miss.R
\name{recode_miss}
\alias{recode_miss}
\title{Recode Values to NA}
\usage{
recode_miss(x, values = NA)
}
\arguments{
\item{x}{A vector.}

\item{values}{The values that you want to recode to NA.}
}
\value{
A vector with desired values recoded to NA.
}
\description{
Recode Values to NA
}
\examples{
# Example data set:
(df <- data.frame(
  edu = c(1, 2, 4, 7, 4),
  marital = c(1, 1, 7, 9, 3),
  race = c(1, 1, 3, 2, NA),
  month = c("May", "Feb", "Sep", "UNK", "?"),
  stringsAsFactors = FALSE
))

# Recode a single numeric variable
recode_miss(df$marital, values = c(7, 9))

# Recode a single character variable
recode_miss(df$month, values = c("UNK", "?"))

# Recode multiple variables
lapply(df, recode_miss, values = c(7, 9, "UNK", "?"))
}
