% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Estimate Means and Related Statistics for Continuous Variables}
\usage{
stats(x, digits = 3)
}
\arguments{
\item{x}{A continuous variable.}

\item{digits}{Rounds the values returned to the specified number of decimal
places (default 3).}
}
\value{
A numeric vector contining the n, mean, median, variance, standard
 deviation, standard error of the mean, t critical variable, 95 percent
 confidnece interval, min, and max.
}
\description{
Estimate Means and Related Statistics for Continuous Variables
}
\examples{
data(mtcars)

# Single univariate analysis
stats(mtcars$mpg)
}
