% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabstat.R
\name{tabstat}
\alias{tabstat}
\title{Compact Table of Summary Statistics}
\usage{
tabstat(x, digits = 3, stats = "default")
}
\arguments{
\item{x}{A continuous variable.}

\item{digits}{Rounds the values returned to the specified number of decimal
places (default 3).}

\item{stats}{Return specified statistics. Options include:
\describe{
  \item{n}{Count of nonmissing values of x}
  \item{nmiss}{Count of missing values of x}
  \item{ci}{95 percent confidence interval for the mean of x}
  \item{sum}{Sum of x}
  \item{max}{Maximum value of x}
  \item{min}{Minimum value of x}
  \item{range}{(Maximum value of x) - (minimum value of x)}
  \item{sd}{Standard deviation of x}
  \item{var}{Variance of x}
  \item{cv}{Coefficient of variation (sd / mean) of x}
  \item{sem}{Standard error of the mean of x}
  \item{skew}{Skewness of x}
  \item{kurt}{Kurtosis of x}
  \item{p1}{1st percentile of x}
  \item{p5}{5th percentile of x}
  \item{p10}{10th percentile of x}
  \item{p25}{25th percentile of x}
  \item{p50}{Median value of x}
  \item{median}{Median value of x}
  \item{p75}{75th percentile of x}
  \item{p90}{90th percentile of x}
  \item{p95}{95th percentile of x}
  \item{p99}{99th percentile of x}
  \item{iqr}{Interquartile range (p75 - p25)}
  \item{q}{Equivalent to specifying p25 p50 p75}
}}
}
\value{
A data frame. By default, the data frame contains the variable name
 and mean.
}
\description{
Based on Stata's "tabstat" command. "tabstat displays summary
 statistics for a series of numeric variables in one table.  It allows you
 to specify the list of statistics to be displayed.  Statistics can be
 calculated (conditioned on) another variable.  tabstat allows substantial
 flexibility in terms of the statistics presented and the format of the
 table" (Stata, 2016).
}
\examples{
data(mtcars)

# Single univariate analysis with Defaults
tabstat(mtcars$mpg)

# Single univariate analysis with all stats
tabstat(mtcars$mpg, stats = c("n", "nmiss", "ci", "sum", "max", "min",
"range", "sd", "var", "cv", "sem", "skew", "kurt", "p1", "p5", "p10",
"p25", "p50", "median", "p75", "p90", "p95", "p99", "iqr", "q"))
}
\references{
Stata 14 help for tabstat
 \url{http://www.stata.com/help.cgi?tabstat}
}
