% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{Format freq_table and mean_table for Publication Tables

#' @description The format_table function takes the results of the
  function or the mean_table function and formats them for publication
  tables. For example, a mean and 95% confidence interval would be formatted
  as 24.00 (21.00 - 27.00) by default.}
\usage{
format_table(.data, digits = 2, ...)
}
\arguments{
\item{.data}{A data frame of class mean_table.}

\item{digits}{Determines the number of decimal place to display. Passed to
the "nsmall =" parameter of the format function.}

\item{...}{Other parameters to be passed on.}
}
\value{
A tibble.
}
\examples{
library(tidyverse)
library(bfuncs)

data(mtcars)

# Overall mean table
mtcars \%>\%
  mean_table(mpg) \%>\%
  format_table()

#> # A tibble: 1 x 2
#>     var               mean_95
#>   <chr>                 <chr>
#> 1   mpg 20.09 (17.92 - 22.26)

# Grouped means table
mtcars \%>\%
  group_by(cyl) \%>\%
  mean_table(mpg) \%>\%
  format_table()

#> # A tibble: 3 x 3
#>     cyl   var               mean_95
#>   <dbl> <chr>                 <chr>
#> 1     4   mpg 26.66 (23.63 - 29.69)
#> 2     6   mpg 19.74 (18.40 - 21.09)
#> 3     8   mpg 15.10 (13.62 - 16.58)
}
