% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_table.R
\name{freq_table}
\alias{freq_table}
\title{Estimate Percents and 95 Percent Confidence Intervals}
\usage{
freq_table(x, t_prob = 0.975, ci_type = "log", ...)
}
\arguments{
\item{x}{A grouped tibble, i.e., class == "grouped_df"}

\item{t_prob}{(1 - alpha / 2). Default value is 0.975, which corresponds to
an alpha of 0.05. Used to calculate a critical value from Student's t
distribution with n - 1 degrees of freedom.}

\item{ci_type}{Selects the method used to estimate 95 percent confidence intervals.
The default for one-way and two-way tables is log transformed ("log"). For
one-way tables only, ci_type can optionally calculate Wald ("linear")
confidence intervals using the "wald" argument.}

\item{...}{Other parameters to be passed on.}
}
\value{
A tibble
}
\description{
The freq_table function produces one-way and two-way frequency
  tables for categorical variables. In addition to frequencies, the
  freq_table function displays percentages, and the standard errors and 95%
  confidence intervals of the percentages. For two-way tables only,
  freq_table also displays row (subgroup) percentages, standard errors,
  and 95 percent confidence intervals.

  freq_table is intended to be used in a dplyr pipeline. Specifically,
  freq_table expects the x argument to be a grouped tibble created with
  dplyr's group_by function.

  All standard errors are calculated as some version of:
  sqrt(proportion * (1 - proportion) / (n - 1))

  For one-way tables, the default 95 percent confidence intervals displayed are
  log transformed confidence intervals equivalent to those used by Stata.
  Additionally, freq_table will return Wald ("linear") confidence intervals
  if the argument to ci_type = "wald".

  For two-way tables, freq_table returns log transformed confidence
  intervals equivalent to those used by Stata.
}
\examples{
library(tidyverse)

data(mtcars)

# One-way frequency table
mtcars \%>\%
  group_by(am) \%>\%
  freq_table()

# Two-way frequency table
mtcars \%>\%
  group_by(am, cyl) \%>\%
  freq_table()

# More examples in r_notes > frequency_tables.Rmd
}
\references{
Agresti, A. (2012). Categorical Data Analysis (3rd ed.). Hoboken, NJ: Wiley.

  SAS documentation: https://support.sas.com/documentation/cdl/en/statug/63347/HTML/default/viewer.htm#statug_surveyfreq_a0000000221.htm

  Stata documentation: https://www.stata.com/manuals13/rproportion.pdf
}
